Renamed Main
Deleted Unused iml files that were added erroneously
Renamed Main.java to TrainModelUserInterface.java
