--- TrainModel/src/trainmodel/Main.java
@@ -1,24 +1,0 @@
-package trainmodel;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import trainmodel.view.TrainModelController;
-
-public class Main extends Application {
-
-  @Override
-  public void start(Stage primaryStage) throws Exception {
-    Parent root = FXMLLoader.load(getClass().getResource("view/TrainModelView.fxml"));
-    primaryStage.setTitle("Train 1");
-    primaryStage.setScene(new Scene(root, 800, 450));
-    primaryStage.show();
-  }
-
-
-  public static void main(String[] args) {
-    launch(args);
-  }
-}
