+++ TrainModel/src/trainmodel/view/trainModelView.fxml
